backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.6
django-allauth==0.55.2
django-environ==0.11.2
djangorestframework @ git+https://github.com/encode/django-rest-framework.git@f56b85b7dd7e4f786e0769bba6b7609d4507da83
djangorestframework-simplejwt==5.3.0
idna==2.10
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.25.1
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.18
